chore(events_processor): create a file for processor functions
... preparing terrain for refactoring and for process tree processors.
